requirements.txt
flask
python_dotenv
langchain
pytesseract
huggingface_hub
openai
opencv-python
# chromadb
# tiktoken

